requirements.txt
# Beauty OS — Python Dependencies

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# LLM providers
google-genai==1.5.0
anthropic==0.40.0
openai==1.58.1

# SMS
twilio==9.4.0

# HTTP client
requests==2.32.3

# Environment variables
python-dotenv==1.0.1

# Scheduler
schedule==1.2.2

# Form parsing (for Twilio webhooks)
python-multipart==0.0.19
